Multi scroller.
Scrolls the screen joystick controlled in 4 directions.
Displays a cursor in the middle of the screen and
plots a char at this pos, everytime the stick is moved.

This prg shows how multi directonal fine scrolling can be done and
how to locate or set a character inside the scrolling area.