Scroller advanced, Improved version of "Scroller.asm"
Scrolls the screen joystick controlled in 4 directions.
Displays a cursor in the middle of the screen and
plots a char at this pos, everytime the stick is moved.

Cursor is not frozen in center of playfield. When reaching the borders
of the playfield, cursor is moved towards the borders allowing it to reach the whole playfield.

Plot routine now can also read char from screen ram (set 'readflag' to a val > 0).

"Disadvantage": Screen always moves unless cursor is near either of the playfeld boders.