Scroller (even) more advanced. Improved version of "Scroller_advanced.asm".
Does the same as 'Scroller_advanced', but moves not the whole playfield
when the cursor is moved and is is not near of either edge of the playfield.

Plot routine now can also a read char from screen ram (set 'readflag' to a val > 0).
When the cursor is moved, char overlapped will be shown in the status line on top of screen.

This source is a good framework to build a playfield editor or a game engine.